+++ sdk/resourcemanager/containerinstance/armcontainerinstance/containergroups_client_example_test.go
@@ -18,7 +18,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/containerinstance/armcontainerinstance/v2"
 )
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/e60df62e9e0d88462e6abba81a76d94eab000f0d/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/stable/2023-05-01/examples/ContainerGroupsList.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/4995036206058b8a75070667995b8feb038549be/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/preview/2024-10-01-preview/examples/ContainerGroupsList.json
 func ExampleContainerGroupsClient_NewListPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -118,7 +118,7 @@ func ExampleContainerGroupsClient_NewListPager() {
 	}
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/e60df62e9e0d88462e6abba81a76d94eab000f0d/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/stable/2023-05-01/examples/ContainerGroupsListByResourceGroup.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/4995036206058b8a75070667995b8feb038549be/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/preview/2024-10-01-preview/examples/ContainerGroupsListByResourceGroup.json
 func ExampleContainerGroupsClient_NewListByResourceGroupPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -208,7 +208,7 @@ func ExampleContainerGroupsClient_NewListByResourceGroupPager() {
 	}
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/e60df62e9e0d88462e6abba81a76d94eab000f0d/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/stable/2023-05-01/examples/ContainerGroupsGetPriority.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/4995036206058b8a75070667995b8feb038549be/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/preview/2024-10-01-preview/examples/ContainerGroupsGetPriority.json
 func ExampleContainerGroupsClient_Get_containerGroupsGetWithPriority() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -304,7 +304,7 @@ func ExampleContainerGroupsClient_Get_containerGroupsGetWithPriority() {
 	// 	}
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/e60df62e9e0d88462e6abba81a76d94eab000f0d/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/stable/2023-05-01/examples/ContainerGroupsGet_Failed.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/4995036206058b8a75070667995b8feb038549be/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/preview/2024-10-01-preview/examples/ContainerGroupsGet_Failed.json
 func ExampleContainerGroupsClient_Get_containerGroupsGetFailed() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -421,7 +421,7 @@ func ExampleContainerGroupsClient_Get_containerGroupsGetFailed() {
 	// }
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/e60df62e9e0d88462e6abba81a76d94eab000f0d/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/stable/2023-05-01/examples/ContainerGroupsGet_Succeeded.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/4995036206058b8a75070667995b8feb038549be/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/preview/2024-10-01-preview/examples/ContainerGroupsGet_Succeeded.json
 func ExampleContainerGroupsClient_Get_containerGroupsGetSucceeded() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -503,7 +503,7 @@ func ExampleContainerGroupsClient_Get_containerGroupsGetSucceeded() {
 	// 	}
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/e60df62e9e0d88462e6abba81a76d94eab000f0d/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/stable/2023-05-01/examples/ContainerGroupCreateConfidential.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/4995036206058b8a75070667995b8feb038549be/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/preview/2024-10-01-preview/examples/ContainerGroupCreateConfidential.json
 func ExampleContainerGroupsClient_BeginCreateOrUpdate_confidentialContainerGroup() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -630,7 +630,7 @@ func ExampleContainerGroupsClient_BeginCreateOrUpdate_confidentialContainerGroup
 	// 	}
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/e60df62e9e0d88462e6abba81a76d94eab000f0d/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/stable/2023-05-01/examples/ContainerGroupExtensions.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/4995036206058b8a75070667995b8feb038549be/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/preview/2024-10-01-preview/examples/ContainerGroupExtensions.json
 func ExampleContainerGroupsClient_BeginCreateOrUpdate_containerGroupCreateWithExtensions() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -777,7 +777,7 @@ func ExampleContainerGroupsClient_BeginCreateOrUpdate_containerGroupCreateWithEx
 	// }
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/e60df62e9e0d88462e6abba81a76d94eab000f0d/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/stable/2023-05-01/examples/ContainerGroupEncryptionProperties.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/4995036206058b8a75070667995b8feb038549be/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/preview/2024-10-01-preview/examples/ContainerGroupEncryptionProperties.json
 func ExampleContainerGroupsClient_BeginCreateOrUpdate_containerGroupWithEncryptionProperties() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -911,7 +911,7 @@ func ExampleContainerGroupsClient_BeginCreateOrUpdate_containerGroupWithEncrypti
 	// }
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/e60df62e9e0d88462e6abba81a76d94eab000f0d/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/stable/2023-05-01/examples/ContainerGroupsCreateOrUpdate.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/4995036206058b8a75070667995b8feb038549be/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/preview/2024-10-01-preview/examples/ContainerGroupsCreateOrUpdate.json
 func ExampleContainerGroupsClient_BeginCreateOrUpdate_containerGroupsCreateOrUpdate() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -1133,7 +1133,303 @@ func ExampleContainerGroupsClient_BeginCreateOrUpdate_containerGroupsCreateOrUpd
 	// 		}
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/e60df62e9e0d88462e6abba81a76d94eab000f0d/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/stable/2023-05-01/examples/ContainerGroupsCreatePriority.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/4995036206058b8a75070667995b8feb038549be/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/preview/2024-10-01-preview/examples/ContainerGroupsCreateOrUpdateSecretReference.json
+func ExampleContainerGroupsClient_BeginCreateOrUpdate_containerGroupsCreateOrUpdateSecretReference() {
+	cred, err := azidentity.NewDefaultAzureCredential(nil)
+	if err != nil {
+		log.Fatalf("failed to obtain a credential: %v", err)
+	}
+	ctx := context.Background()
+	clientFactory, err := armcontainerinstance.NewClientFactory("<subscription-id>", cred, nil)
+	if err != nil {
+		log.Fatalf("failed to create client: %v", err)
+	}
+	poller, err := clientFactory.NewContainerGroupsClient().BeginCreateOrUpdate(ctx, "demo", "demo1", armcontainerinstance.ContainerGroup{
+		Identity: &armcontainerinstance.ContainerGroupIdentity{
+			Type: to.Ptr(armcontainerinstance.ResourceIdentityTypeUserAssigned),
+			UserAssignedIdentities: map[string]*armcontainerinstance.UserAssignedIdentities{
+				"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/myResourceGroup/providers/Microsoft.ManagedIdentity/userAssignedIdentities/identity-name": {},
+			},
+		},
+		Properties: &armcontainerinstance.ContainerGroupPropertiesProperties{
+			Containers: []*armcontainerinstance.Container{
+				{
+					Name: to.Ptr("demo1"),
+					Properties: &armcontainerinstance.ContainerProperties{
+						Command: []*string{},
+						EnvironmentVariables: []*armcontainerinstance.EnvironmentVariable{
+							{
+								Name:                 to.Ptr("envSecret"),
+								SecureValueReference: to.Ptr("envSecretRef"),
+							}},
+						Image: to.Ptr("privateRegistryImage"),
+						Ports: []*armcontainerinstance.ContainerPort{
+							{
+								Port: to.Ptr[int32](80),
+							}},
+						Resources: &armcontainerinstance.ResourceRequirements{
+							Requests: &armcontainerinstance.ResourceRequests{
+								CPU: to.Ptr[float64](1),
+								Gpu: &armcontainerinstance.GpuResource{
+									Count: to.Ptr[int32](1),
+									SKU:   to.Ptr(armcontainerinstance.GpuSKUK80),
+								},
+								MemoryInGB: to.Ptr[float64](1.5),
+							},
+						},
+						VolumeMounts: []*armcontainerinstance.VolumeMount{
+							{
+								Name:      to.Ptr("volume1"),
+								MountPath: to.Ptr("/mnt/volume1"),
+								ReadOnly:  to.Ptr(false),
+							},
+							{
+								Name:      to.Ptr("volume2"),
+								MountPath: to.Ptr("/mnt/volume2"),
+								ReadOnly:  to.Ptr(false),
+							},
+							{
+								Name:      to.Ptr("volume3"),
+								MountPath: to.Ptr("/mnt/volume3"),
+								ReadOnly:  to.Ptr(true),
+							}},
+					},
+				}},
+			Diagnostics: &armcontainerinstance.ContainerGroupDiagnostics{
+				LogAnalytics: &armcontainerinstance.LogAnalytics{
+					LogType: to.Ptr(armcontainerinstance.LogAnalyticsLogTypeContainerInsights),
+					Metadata: map[string]*string{
+						"test-key": to.Ptr("test-metadata-value"),
+					},
+					WorkspaceID:         to.Ptr("workspaceid"),
+					WorkspaceKey:        to.Ptr("workspaceKey"),
+					WorkspaceResourceID: to.Ptr("/subscriptions/00000000-0000-0000-0000-000000000000/resourcegroups/rg/providers/microsoft.operationalinsights/workspaces/workspace"),
+				},
+			},
+			DNSConfig: &armcontainerinstance.DNSConfiguration{
+				NameServers: []*string{
+					to.Ptr("1.1.1.1")},
+				Options:       to.Ptr("ndots:2"),
+				SearchDomains: to.Ptr("cluster.local svc.cluster.local"),
+			},
+			ImageRegistryCredentials: []*armcontainerinstance.ImageRegistryCredential{
+				{
+					PasswordReference: to.Ptr("privateRegistryKeyRef"),
+					Server:            to.Ptr("demoregistry.azurecr.io"),
+					Username:          to.Ptr("registryUserName"),
+				}},
+			IPAddress: &armcontainerinstance.IPAddress{
+				Type:                              to.Ptr(armcontainerinstance.ContainerGroupIPAddressTypePublic),
+				AutoGeneratedDomainNameLabelScope: to.Ptr(armcontainerinstance.DNSNameLabelReusePolicyUnsecure),
+				DNSNameLabel:                      to.Ptr("dnsnamelabel1"),
+				Ports: []*armcontainerinstance.Port{
+					{
+						Port:     to.Ptr[int32](80),
+						Protocol: to.Ptr(armcontainerinstance.ContainerGroupNetworkProtocolTCP),
+					}},
+			},
+			OSType: to.Ptr(armcontainerinstance.OperatingSystemTypesLinux),
+			SecretReferences: []*armcontainerinstance.SecretReference{
+				{
+					Name:               to.Ptr("envSecretRef"),
+					Identity:           to.Ptr("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/myResourceGroup/providers/Microsoft.ManagedIdentity/userAssignedIdentities/identity-name"),
+					SecretReferenceURI: to.Ptr("https://keyvaultname.vault.azure.net/secrets/envSecret"),
+				},
+				{
+					Name:               to.Ptr("accountKeyRef"),
+					Identity:           to.Ptr("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/myResourceGroup/providers/Microsoft.ManagedIdentity/userAssignedIdentities/identity-name"),
+					SecretReferenceURI: to.Ptr("https://keyvaultname.vault.azure.net/secrets/accountKey"),
+				},
+				{
+					Name:               to.Ptr("volumeSecretRef"),
+					Identity:           to.Ptr("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/myResourceGroup/providers/Microsoft.ManagedIdentity/userAssignedIdentities/identity-name"),
+					SecretReferenceURI: to.Ptr("https://keyvaultname.vault.azure.net/secrets/volumeSecret"),
+				},
+				{
+					Name:               to.Ptr("privateRegistryKeyRef"),
+					Identity:           to.Ptr("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/myResourceGroup/providers/Microsoft.ManagedIdentity/userAssignedIdentities/identity-name"),
+					SecretReferenceURI: to.Ptr("https://keyvaultname.vault.azure.net/secrets/privateRegistryKey"),
+				}},
+			SubnetIDs: []*armcontainerinstance.ContainerGroupSubnetID{
+				{
+					ID: to.Ptr("[resourceId('Microsoft.Network/virtualNetworks/subnets', parameters('vnetName'), parameters('subnetName'))]"),
+				}},
+			Volumes: []*armcontainerinstance.Volume{
+				{
+					Name: to.Ptr("volume1"),
+					AzureFile: &armcontainerinstance.AzureFileVolume{
+						ShareName:                  to.Ptr("shareName"),
+						StorageAccountKeyReference: to.Ptr("accountKeyRef"),
+						StorageAccountName:         to.Ptr("accountName"),
+					},
+				},
+				{
+					Name:     to.Ptr("volume2"),
+					EmptyDir: map[string]any{},
+				},
+				{
+					Name: to.Ptr("volume3"),
+					Secret: map[string]*string{
+						"secretKey1": to.Ptr("SecretValue1InBase64"),
+					},
+					SecretReference: map[string]*string{
+						"secretKey2": to.Ptr("volumeSecretRef"),
+					},
+				}},
+		},
+		Location: to.Ptr("west us"),
+	}, nil)
+	if err != nil {
+		log.Fatalf("failed to finish the request: %v", err)
+	}
+	res, err := poller.PollUntilDone(ctx, nil)
+	if err != nil {
+		log.Fatalf("failed to pull the result: %v", err)
+	}
+	// You could use response here. We use blank identifier for just demo purposes.
+	_ = res
+	// If the HTTP response code is 200 as defined in example definition, your response structure would look as follows. Please pay attention that all the values in the output are fake values for just demo purposes.
+	// res.ContainerGroup = armcontainerinstance.ContainerGroup{
+	// 	Identity: &armcontainerinstance.ContainerGroupIdentity{
+	// 		Type: to.Ptr(armcontainerinstance.ResourceIdentityTypeUserAssigned),
+	// 		TenantID: to.Ptr("00000000-0000-0000-0000-000000000000"),
+	// 		UserAssignedIdentities: map[string]*armcontainerinstance.UserAssignedIdentities{
+	// 			"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/myResourceGroup/providers/Microsoft.ManagedIdentity/userAssignedIdentities/identity-name": &armcontainerinstance.UserAssignedIdentities{
+	// 				ClientID: to.Ptr("00000000-0000-0000-0000-000000000000"),
+	// 				PrincipalID: to.Ptr("00000000-0000-0000-0000-000000000000"),
+	// 			},
+	// 		},
+	// 	},
+	// 	Properties: &armcontainerinstance.ContainerGroupPropertiesProperties{
+	// 		Containers: []*armcontainerinstance.Container{
+	// 			{
+	// 				Name: to.Ptr("demo1"),
+	// 				Properties: &armcontainerinstance.ContainerProperties{
+	// 					Command: []*string{
+	// 					},
+	// 					EnvironmentVariables: []*armcontainerinstance.EnvironmentVariable{
+	// 						{
+	// 							Name: to.Ptr("envSecret"),
+	// 							SecureValueReference: to.Ptr("envSecretRef"),
+	// 					}},
+	// 					Image: to.Ptr("privateRegistryImage"),
+	// 					Ports: []*armcontainerinstance.ContainerPort{
+	// 						{
+	// 							Port: to.Ptr[int32](80),
+	// 					}},
+	// 					Resources: &armcontainerinstance.ResourceRequirements{
+	// 						Requests: &armcontainerinstance.ResourceRequests{
+	// 							CPU: to.Ptr[float64](1),
+	// 							Gpu: &armcontainerinstance.GpuResource{
+	// 								Count: to.Ptr[int32](1),
+	// 								SKU: to.Ptr(armcontainerinstance.GpuSKUK80),
+	// 							},
+	// 							MemoryInGB: to.Ptr[float64](1.5),
+	// 						},
+	// 					},
+	// 					VolumeMounts: []*armcontainerinstance.VolumeMount{
+	// 						{
+	// 							Name: to.Ptr("volume1"),
+	// 							MountPath: to.Ptr("/mnt/volume1"),
+	// 							ReadOnly: to.Ptr(false),
+	// 						},
+	// 						{
+	// 							Name: to.Ptr("volume2"),
+	// 							MountPath: to.Ptr("/mnt/volume2"),
+	// 							ReadOnly: to.Ptr(false),
+	// 						},
+	// 						{
+	// 							Name: to.Ptr("volume3"),
+	// 							MountPath: to.Ptr("/mnt/volume3"),
+	// 							ReadOnly: to.Ptr(true),
+	// 					}},
+	// 				},
+	// 		}},
+	// 		Diagnostics: &armcontainerinstance.ContainerGroupDiagnostics{
+	// 			LogAnalytics: &armcontainerinstance.LogAnalytics{
+	// 				WorkspaceID: to.Ptr("workspaceid"),
+	// 			},
+	// 		},
+	// 		DNSConfig: &armcontainerinstance.DNSConfiguration{
+	// 			NameServers: []*string{
+	// 				to.Ptr("1.1.1.1")},
+	// 				Options: to.Ptr("ndots:2"),
+	// 				SearchDomains: to.Ptr("cluster.local svc.cluster.local"),
+	// 			},
+	// 			ImageRegistryCredentials: []*armcontainerinstance.ImageRegistryCredential{
+	// 				{
+	// 					PasswordReference: to.Ptr("privateRegistryKeyRef"),
+	// 					Server: to.Ptr("demoregistry.azurecr.io"),
+	// 					Username: to.Ptr("registryUserName"),
+	// 			}},
+	// 			IPAddress: &armcontainerinstance.IPAddress{
+	// 				Type: to.Ptr(armcontainerinstance.ContainerGroupIPAddressTypePublic),
+	// 				AutoGeneratedDomainNameLabelScope: to.Ptr(armcontainerinstance.DNSNameLabelReusePolicyUnsecure),
+	// 				DNSNameLabel: to.Ptr("dnsnamelabel1"),
+	// 				Fqdn: to.Ptr("dnsnamelabel1.azure-container.io"),
+	// 				IP: to.Ptr("10.0.0.1"),
+	// 				Ports: []*armcontainerinstance.Port{
+	// 					{
+	// 						Port: to.Ptr[int32](80),
+	// 						Protocol: to.Ptr(armcontainerinstance.ContainerGroupNetworkProtocolTCP),
+	// 				}},
+	// 			},
+	// 			OSType: to.Ptr(armcontainerinstance.OperatingSystemTypesLinux),
+	// 			ProvisioningState: to.Ptr("Succeeded"),
+	// 			SecretReferences: []*armcontainerinstance.SecretReference{
+	// 				{
+	// 					Name: to.Ptr("envSecretRef"),
+	// 					Identity: to.Ptr("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/myResourceGroup/providers/Microsoft.ManagedIdentity/userAssignedIdentities/identity-name"),
+	// 					SecretReferenceURI: to.Ptr("https://keyvaultname.vault.azure.net/secrets/envSecret"),
+	// 				},
+	// 				{
+	// 					Name: to.Ptr("accountKeyRef"),
+	// 					Identity: to.Ptr("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/myResourceGroup/providers/Microsoft.ManagedIdentity/userAssignedIdentities/identity-name"),
+	// 					SecretReferenceURI: to.Ptr("https://keyvaultname.vault.azure.net/secrets/accountKey"),
+	// 				},
+	// 				{
+	// 					Name: to.Ptr("volumeSecretRef"),
+	// 					Identity: to.Ptr("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/myResourceGroup/providers/Microsoft.ManagedIdentity/userAssignedIdentities/identity-name"),
+	// 					SecretReferenceURI: to.Ptr("https://keyvaultname.vault.azure.net/secrets/volumeSecret"),
+	// 				},
+	// 				{
+	// 					Name: to.Ptr("privateRegistryKeyRef"),
+	// 					Identity: to.Ptr("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/myResourceGroup/providers/Microsoft.ManagedIdentity/userAssignedIdentities/identity-name"),
+	// 					SecretReferenceURI: to.Ptr("https://keyvaultname.vault.azure.net/secrets/privateRegistryKey"),
+	// 			}},
+	// 			Volumes: []*armcontainerinstance.Volume{
+	// 				{
+	// 					Name: to.Ptr("volume1"),
+	// 					AzureFile: &armcontainerinstance.AzureFileVolume{
+	// 						ShareName: to.Ptr("shareName"),
+	// 						StorageAccountKeyReference: to.Ptr("accountKeyRef"),
+	// 						StorageAccountName: to.Ptr("accountName"),
+	// 					},
+	// 				},
+	// 				{
+	// 					Name: to.Ptr("volume2"),
+	// 					EmptyDir: map[string]any{
+	// 					},
+	// 				},
+	// 				{
+	// 					Name: to.Ptr("volume3"),
+	// 					Secret: map[string]*string{
+	// 					},
+	// 					SecretReference: map[string]*string{
+	// 						"secretKey2": to.Ptr("volumeSecretRef"),
+	// 					},
+	// 			}},
+	// 		},
+	// 		Name: to.Ptr("demo1"),
+	// 		Type: to.Ptr("Microsoft.ContainerInstance/containerGroups"),
+	// 		ID: to.Ptr("/subscriptions/subid/resourceGroups/demo/providers/Microsoft.ContainerInstance/containerGroups/demo1"),
+	// 		Location: to.Ptr("WestUs"),
+	// 		Zones: []*string{
+	// 			to.Ptr("1")},
+	// 		}
+}
+
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/4995036206058b8a75070667995b8feb038549be/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/preview/2024-10-01-preview/examples/ContainerGroupsCreatePriority.json
 func ExampleContainerGroupsClient_BeginCreateOrUpdate_containerGroupsCreateWithPriority() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -1223,7 +1519,7 @@ func ExampleContainerGroupsClient_BeginCreateOrUpdate_containerGroupsCreateWithP
 	// 	}
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/e60df62e9e0d88462e6abba81a76d94eab000f0d/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/stable/2023-05-01/examples/ContainerGroupsUpdate.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/4995036206058b8a75070667995b8feb038549be/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/preview/2024-10-01-preview/examples/ContainerGroupsUpdate.json
 func ExampleContainerGroupsClient_Update() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -1330,7 +1626,7 @@ func ExampleContainerGroupsClient_Update() {
 	// }
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/e60df62e9e0d88462e6abba81a76d94eab000f0d/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/stable/2023-05-01/examples/ContainerGroupsDelete.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/4995036206058b8a75070667995b8feb038549be/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/preview/2024-10-01-preview/examples/ContainerGroupsDelete.json
 func ExampleContainerGroupsClient_BeginDelete() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -1416,7 +1712,7 @@ func ExampleContainerGroupsClient_BeginDelete() {
 	// 	}
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/e60df62e9e0d88462e6abba81a76d94eab000f0d/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/stable/2023-05-01/examples/ContainerGroupsRestart.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/4995036206058b8a75070667995b8feb038549be/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/preview/2024-10-01-preview/examples/ContainerGroupsRestart.json
 func ExampleContainerGroupsClient_BeginRestart() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -1437,7 +1733,7 @@ func ExampleContainerGroupsClient_BeginRestart() {
 	}
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/e60df62e9e0d88462e6abba81a76d94eab000f0d/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/stable/2023-05-01/examples/ContainerGroupsStop.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/4995036206058b8a75070667995b8feb038549be/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/preview/2024-10-01-preview/examples/ContainerGroupsStop.json
 func ExampleContainerGroupsClient_Stop() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -1454,7 +1750,7 @@ func ExampleContainerGroupsClient_Stop() {
 	}
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/e60df62e9e0d88462e6abba81a76d94eab000f0d/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/stable/2023-05-01/examples/ContainerGroupsStart.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/4995036206058b8a75070667995b8feb038549be/specification/containerinstance/resource-manager/Microsoft.ContainerInstance/preview/2024-10-01-preview/examples/ContainerGroupsStart.json
 func ExampleContainerGroupsClient_BeginStart() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {